refactor: extract update UI from App
